backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
sqlalchemy = "^2.0.23"
paho-mqtt = "^1.6.1"
uvicorn = "^0.25.0"
oauthlib = "^3.2.2"
celery = "^5.3.6"
redis = "^5.0.1"
python-dotenv="^1.0.0"
pandas="^2.1.4"
psycopg2 = "^2.9.9"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
